requirements.txt
# Core
pyspark==3.4.1
kafka-python==2.0.2
redis==5.0.1
snowflake-connector-python==3.6.0
sqlalchemy==2.0.23
pyiceberg==0.5.1

# Data processing
pandas==2.1.4
numpy==1.26.2
pyarrow==14.0.1

# Schema validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Configuration
python-dotenv==1.0.0
PyYAML==6.0.1

# Orchestration
apache-airflow==2.8.0

# dbt (Snowflake adapter)
dbt-snowflake==1.7.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
freezegun==1.3.1
fakeredis==2.20.1

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0

# Development
black==23.12.1
ruff==0.1.9
mypy==1.8.0
